backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==23.4
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
